Moved the folder from hubble to hubblestack
