+++ hubblestack/extmods/modules/win_pulsar.py
