+++ hubblestack/files/hubblestack_nova/win_pkg.py
