+++ hubblestack/files/hubblestack_nova/win_reg.py
