+++ hubblestack/files/hubblestack_nova/win_secedit.py
